requirements.txt
flask==3.0.3
gunicorn==22.0.0
psycopg2-binary
pgvector==0.4.2
tiktoken==0.9.0
python-dotenv==1.0.1
